Implement Initial Scaffolder
This is an initial implementation of a scaffolding tool that can be used to scaffold Fusion projects, plugins and potentially other items in the future. The expectation is that this will stay fairly generic, but consuming libraries (e.g. `create-fusion-app` or `create-fusion-plugin`) would include more specific CLIs.

For now, this library has a basic implementation: It takes a path to a template and a project name and copies the files in the template into a new directory, and allows for interpolation of values into files in the template via nunjucks when using a `.njk` extension.

Not yet implemented is a way to gather options/values to be passed to the nunjucks compiler through an `index.js` file at the root of the template. In addition, we may want to do dependency installation as part of the scaffolder and possibly more...