Add index.js context functionality
This adds functionality such that an `index.js` file in the root of the temp
latewill be called prior to `.njk` compilation to gather additional context/valu
es for interpolation. For example, a template author could request additional da
ta from the user (via stdin/stdout) in the index.js file and then use that data
in files that pass through nunjucks compilation.